requirements.txt
PyYAML==6.0.1
boto3==1.28.15